Fixing issues with accessing configuration
Plugin API of database2 has changed for getConf() isn't returning `null` but `false` on missing some configuration option anymore. This change is due to supporting provision of default value by caller. If this provision is missing, `false` is used as default. The fix is ensuring to use `null` as default.

See issue #5.